requirements_macos.txt
contourpy==1.2.1
cycler==0.12.1
filelock==3.18.0
fonttools==4.58.4
fsspec==2025.5.1
Jinja2==3.1.6
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.8.4
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
packaging==25.0
pillow==11.2.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
six==1.17.0
sympy==1.14.0
torch
torchvision
typing_extensions==4.14.0
tqdm
torcheval
scikit-learn
plantcv